# src/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: LiftLog API
  version: 1.0.0
servers:
  - url: /
    description: Same-origin (Swagger UI served from this API)
  - url: https://loadlog-api.onrender.com
    description: Render production
  - url: http://localhost:3000
    description: Local development (localhost)
  - url: http://127.0.0.1:3000
    description: Local development (loopback)
paths:
...
  /ai/coach-chat:
    post:
      operationId: coachChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachChatRequest'
      responses:
        '200':
          description: Coach chatbot response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoachChatResponse'
components:
  schemas:
...
    CoachChatRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          minLength: 1
        userId:
          type: string
        history:
          type: array
          maxItems: 20
          items:
            $ref: '#/components/schemas/CoachChatMessage'
        profile:
          $ref: '#/components/schemas/CoachProfile'
    CoachChatMessage:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          minLength: 1
    CoachProfile:
      type: object
      properties:
        goal:
          type: string
        dietaryPreferences:
          type: string
        injuriesOrLimitations:
          type: string
        experienceLevel:
          type: string
          enum: [beginner, intermediate, advanced]
...
    CoachChatResponse:
      type: object
      required:
        - reply
        - model
      properties:
        reply:
          type: string
        model:
          type: string